refactor(cluster): split distributor runtime distribution tests
