+++ src/agilab/core/test/test_agi_distributor.py
@@ -749,18 +749,6 @@ def test_is_local():
     assert not AgiEnv.is_local("8.8.8.8"), "8.8.8.8 should not be considered local."
 
 
-def test_hardware_supports_rapids_true_and_false(monkeypatch):
-    monkeypatch.setattr(agi_distributor_module.subprocess, "run", lambda *_a, **_k: None)
-    assert AGI._hardware_supports_rapids() is True
-
-    monkeypatch.setattr(
-        agi_distributor_module.subprocess,
-        "run",
-        lambda *_a, **_k: (_ for _ in ()).throw(FileNotFoundError("nvidia-smi missing")),
-    )
-    assert AGI._hardware_supports_rapids() is False
-
-
 @pytest.mark.asyncio
 async def test_agi_run_delegates_to_benchmark_with_sorted_mode_list(monkeypatch):
     env = AgiEnv(apps_path=Path("src/agilab/apps/builtin"), app="mycode_project", verbose=0)
@@ -1227,169 +1215,6 @@ async def test_agi_run_requires_base_worker_cls():
         )
 
 
-@pytest.mark.asyncio
-async def test_start_scheduler_local_switches_port_and_connects(monkeypatch, tmp_path):
-    cluster_pck = tmp_path / "cluster"
-    (cluster_pck / "agi_distributor").mkdir(parents=True, exist_ok=True)
-    (cluster_pck / "agi_distributor" / "cli.py").write_text("print('cli')", encoding="utf-8")
-    active_app = tmp_path / "app"
-    active_app.mkdir(parents=True, exist_ok=True)
-    (active_app / "pyproject.toml").write_text("[project]\nname='app'\n", encoding="utf-8")
-
-    AGI._mode = AGI.DASK_MODE
-    AGI._mode_auto = False
-    AGI._workers = {"127.0.0.1": 1}
-    AGI._worker_init_error = False
-    AGI.env = SimpleNamespace(
-        wenv_rel=Path("wenv"),
-        wenv_abs=tmp_path / "wenv",
-        active_app=active_app,
-        app="demo_app",
-        uv="uv",
-        envars={},
-        cluster_pck=cluster_pck,
-        export_local_bin="",
-        is_local=lambda ip: ip == "127.0.0.1",
-    )
-    AGI.env.wenv_abs.mkdir(parents=True, exist_ok=True)
-    calls = {"bg": [], "set_env": []}
-
-    async def _fake_send(*_args, **_kwargs):
-        return None
-
-    async def _fake_kill(*_args, **_kwargs):
-        return None
-
-    async def _fake_connect(*_args, **_kwargs):
-        return "fake-client"
-
-    async def _fake_detect(_ip):
-        return "export PATH=\"$HOME/.local/bin:$PATH\"; "
-
-    async def _fake_sleep(_delay):
-        return None
-
-    async def _fake_port_release(*_args, **_kwargs):
-        return False
-
-    monkeypatch.setattr(
-        AGI,
-        "_get_scheduler",
-        staticmethod(lambda scheduler: ("127.0.0.1", 8786) if scheduler else ("127.0.0.1", 8786)),
-    )
-    monkeypatch.setattr(AGI, "send_file", staticmethod(_fake_send))
-    monkeypatch.setattr(AGI, "_kill", staticmethod(_fake_kill))
-    monkeypatch.setattr(AGI, "_wait_for_port_release", staticmethod(_fake_port_release))
-    monkeypatch.setattr(AGI, "find_free_port", staticmethod(lambda *_a, **_k: 8899))
-    monkeypatch.setattr(AGI, "_detect_export_cmd", staticmethod(_fake_detect))
-    monkeypatch.setattr(AGI, "_connect_scheduler_with_retry", staticmethod(_fake_connect))
-    monkeypatch.setattr(agi_distributor_module.asyncio, "sleep", _fake_sleep)
-    monkeypatch.setattr(
-        agi_distributor_module.AgiEnv,
-        "set_env_var",
-        staticmethod(lambda *args: calls["set_env"].append(args)),
-    )
-    monkeypatch.setattr(
-        AGI,
-        "_exec_bg",
-        staticmethod(lambda cmd, cwd: calls["bg"].append((cmd, cwd))),
-    )
-
-    ok = await AGI._start_scheduler("127.0.0.1")
-
-    assert ok is True
-    assert AGI._scheduler_port == 8899
-    assert AGI._dask_client == "fake-client"
-    assert AGI._install_done is True
-    assert calls["bg"]
-    assert any(entry[0] == "127.0.0.1_CMD_PREFIX" for entry in calls["set_env"])
-
-
-@pytest.mark.asyncio
-async def test_connect_scheduler_with_retry_succeeds_after_retry(monkeypatch):
-    attempts = {"n": 0}
-
-    async def _fake_client(_address, heartbeat_interval=5000, timeout=1.0):
-        attempts["n"] += 1
-        if attempts["n"] < 3:
-            raise RuntimeError("not ready")
-        return {"connected": True, "timeout": timeout, "heartbeat": heartbeat_interval}
-
-    async def _fake_sleep(_delay):
-        return None
-
-    monkeypatch.setattr(agi_distributor_module, "Client", _fake_client)
-    monkeypatch.setattr(agi_distributor_module.asyncio, "sleep", _fake_sleep)
-
-    client = await AGI._connect_scheduler_with_retry("tcp://127.0.0.1:8786", timeout=2.0)
-    assert client["connected"] is True
-    assert attempts["n"] == 3
-
-
-@pytest.mark.asyncio
-async def test_connect_scheduler_with_retry_times_out(monkeypatch):
-    async def _fake_client(_address, heartbeat_interval=5000, timeout=1.0):
-        raise RuntimeError("never ready")
-
-    async def _fake_sleep(_delay):
-        return None
-
-    clock = {"t": 0.0}
-
-    def _monotonic():
-        clock["t"] += 0.25
-        return clock["t"]
-
-    monkeypatch.setattr(agi_distributor_module, "Client", _fake_client)
-    monkeypatch.setattr(agi_distributor_module.asyncio, "sleep", _fake_sleep)
-    monkeypatch.setattr(agi_distributor_module.time, "monotonic", _monotonic)
-
-    with pytest.raises(RuntimeError, match="Failed to instantiate Dask Client"):
-        await AGI._connect_scheduler_with_retry("tcp://127.0.0.1:8786", timeout=0.1)
-
-
-@pytest.mark.asyncio
-async def test_detect_export_cmd_local_and_remote(monkeypatch):
-    monkeypatch.setattr(AgiEnv, "is_local", staticmethod(lambda ip: ip == "127.0.0.1"))
-    had_attr = hasattr(AgiEnv, "export_local_bin")
-    old_export = getattr(AgiEnv, "export_local_bin", None)
-    setattr(AgiEnv, "export_local_bin", "LOCAL_PREFIX ")
-    assert await AGI._detect_export_cmd("127.0.0.1") == "LOCAL_PREFIX "
-
-    async def _fake_exec(_ip, _cmd):
-        return "Linux"
-
-    monkeypatch.setattr(AGI, "exec_ssh", staticmethod(_fake_exec))
-    assert await AGI._detect_export_cmd("10.0.0.2") == 'export PATH="$HOME/.local/bin:$PATH";'
-    if had_attr:
-        setattr(AgiEnv, "export_local_bin", old_export)
-    else:
-        delattr(AgiEnv, "export_local_bin")
-
-
-@pytest.mark.asyncio
-async def test_detect_export_cmd_returns_empty_for_non_posix(monkeypatch):
-    monkeypatch.setattr(AgiEnv, "is_local", staticmethod(lambda _ip: False))
-
-    async def _fake_exec(_ip, _cmd):
-        return "Windows_NT"
-
-    monkeypatch.setattr(AGI, "exec_ssh", staticmethod(_fake_exec))
-    assert await AGI._detect_export_cmd("10.0.0.3") == ""
-
-
-def test_dask_env_prefix_and_scale_cluster_trim_workers():
-    AGI._dask_log_level = ""
-    assert AGI._dask_env_prefix() == ""
-    AGI._dask_log_level = "INFO"
-    assert "DASK_DISTRIBUTED__LOGGING__distributed=INFO" in AGI._dask_env_prefix()
-
-    AGI._workers = {"10.0.0.1": 1}
-    AGI._dask_workers = ["10.0.0.1:1001", "10.0.0.1:1002", "10.0.0.2:1001"]
-    AGI._scale_cluster()
-    assert AGI._dask_workers == ["10.0.0.1:1001"]
-
-
 def test_exec_bg_raises_when_background_job_fails():
     class _Jobs:
         def __init__(self):
@@ -1503,444 +1328,6 @@ async def test_deploy_remote_worker_non_source_flow(monkeypatch, tmp_path):
 
 
 @pytest.mark.asyncio
-async def test_start_launches_workers_and_uploads_eggs(monkeypatch, tmp_path):
-    wenv_abs = tmp_path / "worker_env"
-    (wenv_abs / "dist").mkdir(parents=True, exist_ok=True)
-    (wenv_abs / "dist" / "demo.egg").write_text("x", encoding="utf-8")
-
-    AGI.env = SimpleNamespace(
-        is_local=lambda ip: ip == "127.0.0.1",
-        envars={},
-        uv="uv",
-        wenv_abs=wenv_abs,
-        wenv_rel=Path("worker_env"),
-    )
-    AGI._mode = AGI.DASK_MODE
-    AGI._mode_auto = False
-    AGI._workers = {"127.0.0.1": 1, "10.0.0.2": 1}
-    AGI._scheduler = "127.0.0.1:8786"
-    AGI._worker_init_error = False
-    calls = {"bg": [], "remote": [], "uploaded": []}
-
-    class _Client:
-        def upload_file(self, path):
-            calls["uploaded"].append(path)
-
-    async def _fake_start_scheduler(_scheduler):
-        return True
-
-    async def _fake_detect(_ip):
-        return "export PATH=\"$HOME/.local/bin:$PATH\"; "
-
-    async def _fake_sync(timeout=60):
-        return None
-
-    async def _fake_build_remote():
-        return None
-
-    async def _fake_exec_ssh_async(ip, cmd):
-        calls["remote"].append((ip, cmd))
-        return ""
-
-    monkeypatch.setattr(AGI, "_dask_client", _Client())
-    monkeypatch.setattr(AGI, "_start_scheduler", staticmethod(_fake_start_scheduler))
-    monkeypatch.setattr(AGI, "_detect_export_cmd", staticmethod(_fake_detect))
-    monkeypatch.setattr(AGI, "_sync", staticmethod(_fake_sync))
-    monkeypatch.setattr(AGI, "_build_lib_remote", staticmethod(_fake_build_remote))
-    monkeypatch.setattr(AGI, "exec_ssh_async", staticmethod(_fake_exec_ssh_async))
-    monkeypatch.setattr(
-        agi_distributor_module.AgiEnv,
-        "set_env_var",
-        staticmethod(lambda *_a, **_k: None),
-    )
-    monkeypatch.setattr(
-        AGI,
-        "_exec_bg",
-        staticmethod(lambda cmd, cwd: calls["bg"].append((cmd, cwd))),
-    )
-
-    await AGI._start("127.0.0.1")
-    await asyncio.sleep(0)
-
-    assert calls["bg"]
-    assert any(ip == "10.0.0.2" for ip, _ in calls["remote"])
-    assert calls["uploaded"]
-
-
-@pytest.mark.asyncio
-async def test_stop_retires_workers_and_shutdown(monkeypatch):
-    class _Client:
-        def __init__(self):
-            self.info_calls = 0
-            self.retire_calls = 0
-            self.shutdown_calls = 0
-
-        async def scheduler_info(self):
-            self.info_calls += 1
-            if self.info_calls == 1:
-                return {"workers": {"tcp://127.0.0.1:8787": {}}}
-            return {"workers": {}}
-
-        async def retire_workers(self, workers, close_workers=True, remove=True):
-            self.retire_calls += 1
-
-        async def shutdown(self):
-            self.shutdown_calls += 1
-
-    AGI._dask_client = _Client()
-    AGI._mode_auto = False
-    AGI._mode = AGI.DASK_MODE
-    AGI._TIMEOUT = 3
-    AGI.env = SimpleNamespace()
-    closed = {"count": 0}
-
-    async def _fake_close_all():
-        closed["count"] += 1
-
-    async def _fake_sleep(_delay):
-        return None
-
-    monkeypatch.setattr(AGI, "_close_all_connections", staticmethod(_fake_close_all))
-    monkeypatch.setattr(agi_distributor_module.asyncio, "sleep", _fake_sleep)
-
-    await AGI._stop()
-
-    assert AGI._dask_client.retire_calls >= 1
-    assert AGI._dask_client.shutdown_calls == 1
-    assert closed["count"] == 1
-
-
-@pytest.mark.asyncio
-async def test_run_raises_when_worker_venv_is_missing(tmp_path, monkeypatch):
-    AGI.env = SimpleNamespace(
-        envars={},
-        wenv_abs=tmp_path / "missing_wenv",
-        debug=False,
-        verbose=0,
-        uv="uv",
-        target_worker="demo_worker",
-    )
-    AGI._mode = AGI.PYTHON_MODE
-    AGI._workers = {"127.0.0.1": 1}
-    AGI._args = {}
-    monkeypatch.chdir(tmp_path)
-    with pytest.raises(FileNotFoundError, match="Worker installation"):
-        await AGI._run()
-
-
-@pytest.mark.asyncio
-async def test_run_raises_when_worker_uv_sources_are_stale(tmp_path, monkeypatch):
-    wenv_abs = tmp_path / "wenv"
-    (wenv_abs / ".venv").mkdir(parents=True, exist_ok=True)
-    (wenv_abs / "pyproject.toml").write_text(
-        """
-[tool.uv.sources]
-ilp_worker = { path = "../../PycharmProjects/thales_agilab/apps/ilp_project/src/ilp_worker" }
-""".strip(),
-        encoding="utf-8",
-    )
-    AGI.env = SimpleNamespace(
-        envars={},
-        wenv_abs=wenv_abs,
-        debug=False,
-        verbose=0,
-        uv="uv",
-        target_worker="demo_worker",
-    )
-    AGI._mode = AGI.PYTHON_MODE
-    AGI._workers = {"127.0.0.1": 1}
-    AGI._args = {}
-    monkeypatch.chdir(tmp_path)
-
-    with pytest.raises(RuntimeError, match="stale or incomplete"):
-        await AGI._run()
-
-
-@pytest.mark.asyncio
-async def test_run_debug_branch_returns_list_result(tmp_path, monkeypatch):
-    wenv_abs = tmp_path / "wenv"
-    (wenv_abs / ".venv").mkdir(parents=True, exist_ok=True)
-    AGI.env = SimpleNamespace(
-        envars={},
-        wenv_abs=wenv_abs,
-        debug=True,
-        verbose=1,
-        uv="uv",
-        target_worker="demo_worker",
-    )
-    AGI._mode = AGI.DASK_MODE
-    AGI._workers = {"127.0.0.1": 1}
-    AGI._args = {"alpha": 1}
-    monkeypatch.chdir(tmp_path)
-    calls = {"new": 0, "run": 0, "kill": 0}
-
-    def _fake_new(*_args, **_kwargs):
-        calls["new"] += 1
-
-    async def _fake_run(*_args, **_kwargs):
-        calls["run"] += 1
-        return ["ok", "done"]
-
-    async def _fake_kill(*_args, **_kwargs):
-        calls["kill"] += 1
-        return None
-
-    monkeypatch.setattr(BaseWorker, "_new", staticmethod(_fake_new))
-    monkeypatch.setattr(BaseWorker, "_run", staticmethod(_fake_run))
-    monkeypatch.setattr(AGI, "_kill", staticmethod(_fake_kill))
-
-    result = await AGI._run()
-    assert result == ["ok", "done"]
-    assert calls["new"] == 1
-    assert calls["run"] == 1
-    assert calls["kill"] == 1
-    assert (tmp_path / "dask_worker_0.pid").exists()
-
-
-@pytest.mark.asyncio
-async def test_run_non_debug_branch_parses_last_line(tmp_path, monkeypatch):
-    wenv_abs = tmp_path / "wenv"
-    (wenv_abs / ".venv").mkdir(parents=True, exist_ok=True)
-    AGI.env = SimpleNamespace(
-        envars={},
-        wenv_abs=wenv_abs,
-        debug=False,
-        verbose=0,
-        uv="uv",
-        target_worker="demo_worker",
-    )
-    AGI._mode = AGI.PYTHON_MODE
-    AGI._workers = {"127.0.0.1": 1}
-    AGI._args = {"beta": 2}
-    monkeypatch.chdir(tmp_path)
-
-    async def _fake_kill(*_args, **_kwargs):
-        return None
-
-    async def _fake_run_async(_cmd, _cwd):
-        return "header\nresult-line\n"
-
-    monkeypatch.setattr(AGI, "_kill", staticmethod(_fake_kill))
-    monkeypatch.setattr(agi_distributor_module.AgiEnv, "run_async", staticmethod(_fake_run_async))
-
-    result = await AGI._run()
-    assert result == "result-line"
-
-
-@pytest.mark.asyncio
-async def test_distribute_executes_new_calibration_and_works(monkeypatch):
-    class _Client:
-        def __init__(self):
-            self._gather_calls = 0
-            self.submissions = []
-
-        def scheduler_info(self):
-            return {
-                "workers": {
-                    "tcp://127.0.0.1:8787": {},
-                    "tcp://10.0.0.2:8788": {},
-                }
-            }
-
-        def submit(self, fn, *args, **kwargs):
-            self.submissions.append(getattr(fn, "__name__", str(fn)))
-            return {"fn": getattr(fn, "__name__", "fn"), "args": args, "kwargs": kwargs}
-
-        def gather(self, futures):
-            self._gather_calls += 1
-            if self._gather_calls == 1:
-                return [None for _ in futures]
-            return ["log-a", "log-b"]
-
-    AGI.env = SimpleNamespace(
-        debug=False,
-        target_worker="demo_worker",
-        mode2str=lambda _mode: "dask",
-    )
-    AGI._dask_client = _Client()
-    AGI._workers = {"127.0.0.1": 1, "10.0.0.2": 1}
-    AGI._args = {"k": "v"}
-    AGI._mode = AGI.DASK_MODE
-    AGI.verbose = 0
-    AGI.debug = False
-    called = {"calibration": 0}
-
-    async def _fake_distrib(_env, workers, _args):
-        return workers, [["step-a"], ["step-b"]], [[{"m": 1}], [{"m": 2}]]
-
-    async def _fake_calibration():
-        called["calibration"] += 1
-        AGI._capacity = {"127.0.0.1:8787": 1.0, "10.0.0.2:8788": 1.0}
-
-    monkeypatch.setattr(agi_distributor_module.WorkDispatcher, "_do_distrib", staticmethod(_fake_distrib))
-    monkeypatch.setattr(AGI, "_calibration", staticmethod(_fake_calibration))
-
-    result = await AGI._distribute()
-    assert result.startswith("dask ")
-    assert called["calibration"] == 1
-    assert AGI._work_plan == [["step-a"], ["step-b"]]
-    assert AGI._work_plan_metadata == [[{"m": 1}], [{"m": 2}]]
-    assert "BaseWorker._new" not in AGI._dask_client.submissions
-
-
-@pytest.mark.asyncio
-async def test_sync_waits_until_expected_workers(monkeypatch):
-    class _FakeClient:
-        def __init__(self):
-            self.calls = 0
-
-        def scheduler_info(self):
-            self.calls += 1
-            if self.calls == 1:
-                return {"workers": None}
-            if self.calls == 2:
-                return {"workers": {"tcp://127.0.0.1:8787": {}}}
-            return {
-                "workers": {
-                    "tcp://127.0.0.1:8787": {},
-                    "tcp://10.0.0.2:8788": {},
-                }
-            }
-
-    AGI._workers = {"127.0.0.1": 1, "10.0.0.2": 1}
-    fake_client = _FakeClient()
-    monkeypatch.setattr(agi_distributor_module, "Client", _FakeClient)
-    AGI._dask_client = fake_client
-
-    async def _fake_sleep(_delay):
-        return None
-
-    monkeypatch.setattr(agi_distributor_module.asyncio, "sleep", _fake_sleep)
-
-    await AGI._sync(timeout=2)
-    assert fake_client.calls >= 3
-
-
-@pytest.mark.asyncio
-async def test_sync_raises_timeout_on_repeated_failures(monkeypatch):
-    class _FakeClient:
-        def scheduler_info(self):
-            raise RuntimeError("scheduler down")
-
-    AGI._workers = {"127.0.0.1": 1}
-    fake_client = _FakeClient()
-    monkeypatch.setattr(agi_distributor_module, "Client", _FakeClient)
-    AGI._dask_client = fake_client
-
-    async def _fake_sleep(_delay):
-        return None
-
-    clock = {"t": 0.0}
-
-    def _fake_time():
-        clock["t"] += 0.3
-        return clock["t"]
-
-    monkeypatch.setattr(agi_distributor_module.asyncio, "sleep", _fake_sleep)
-    monkeypatch.setattr(agi_distributor_module.time, "time", _fake_time)
-
-    with pytest.raises(TimeoutError, match="Timeout waiting for all workers"):
-        await AGI._sync(timeout=0.5)
-
-
-@pytest.mark.asyncio
-async def test_main_branches_simulate_install_dask_and_local(monkeypatch):
-    class _Jobs:
-        def flush(self):
-            return None
-
-    monkeypatch.setattr(agi_distributor_module.bg, "BackgroundJobManager", lambda: _Jobs())
-    calls = []
-
-    async def _fake_run():
-        calls.append("run")
-        return "run-result"
-
-    async def _fake_prepare_local():
-        calls.append("prepare_local")
-        return None
-
-    async def _fake_prepare_cluster(_scheduler):
-        calls.append("prepare_cluster")
-        return None
-
-    async def _fake_deploy(_scheduler):
-        calls.append("deploy")
-        return None
-
-    async def _fake_start(_scheduler):
-        calls.append("start")
-        return None
-
-    async def _fake_distribute():
-        calls.append("distribute")
-        return "dist-result"
-
-    async def _fake_stop():
-        calls.append("stop")
-        return None
-
-    monkeypatch.setattr(AGI, "_run", staticmethod(_fake_run))
-    monkeypatch.setattr(AGI, "_prepare_local_env", staticmethod(_fake_prepare_local))
-    monkeypatch.setattr(AGI, "_prepare_cluster_env", staticmethod(_fake_prepare_cluster))
-    monkeypatch.setattr(AGI, "_deploy_application", staticmethod(_fake_deploy))
-    monkeypatch.setattr(AGI, "_start", staticmethod(_fake_start))
-    monkeypatch.setattr(AGI, "_distribute", staticmethod(_fake_distribute))
-    monkeypatch.setattr(AGI, "_stop", staticmethod(_fake_stop))
-    monkeypatch.setattr(AGI, "_update_capacity", staticmethod(lambda: calls.append("update_capacity")))
-    monkeypatch.setattr(AGI, "_clean_dirs_local", staticmethod(lambda: calls.append("clean_dirs_local")))
-    monkeypatch.setattr(AGI, "_clean_job", staticmethod(lambda cond: calls.append(("clean_job", cond))))
-
-    AGI._mode = AGI._SIMULATE_MODE
-    result = await AGI._main("127.0.0.1")
-    assert result == "run-result"
-
-    times = iter([10.0, 14.5])
-    monkeypatch.setattr(agi_distributor_module.time, "time", lambda: next(times))
-    AGI._mode = AGI._INSTALL_MODE | AGI.DASK_MODE
-    result = await AGI._main("127.0.0.1")
-    assert result == 4.5
-
-    AGI._mode = AGI.DASK_MODE
-    result = await AGI._main("127.0.0.1")
-    assert result == "dist-result"
-
-    AGI._mode = AGI.PYTHON_MODE
-    result = await AGI._main("127.0.0.1")
-    assert result == "run-result"
-
-    assert "prepare_local" in calls
-    assert "prepare_cluster" in calls
-    assert "deploy" in calls
-    assert "start" in calls
-    assert "distribute" in calls
-    assert "stop" in calls
-
-
-def test_clean_job_respects_cond_and_verbosity(monkeypatch):
-    class _Jobs:
-        def __init__(self):
-            self.flush_calls = 0
-
-        def flush(self):
-            self.flush_calls += 1
-
-    jobs = _Jobs()
-    AGI._jobs = jobs
-
-    AGI.verbose = 1
-    AGI._clean_job(True)
-    assert jobs.flush_calls == 1
-
-    AGI.verbose = 0
-    AGI._clean_job(True)
-    assert jobs.flush_calls == 2
-
-    AGI._clean_job(False)
-    assert jobs.flush_calls == 2
-
-
-@pytest.mark.asyncio
 async def test_calibration_computes_normalized_capacity(monkeypatch):
     class _Predictor:
         def predict(self, _data):
@@ -2027,62 +1414,6 @@ def test_update_capacity_success_and_guard_paths(tmp_path, monkeypatch):
         AGI._update_capacity()
 
 
-@pytest.mark.asyncio
-async def test_start_scheduler_wraps_non_runtime_error(monkeypatch, tmp_path):
-    app_path = tmp_path / "app"
-    app_path.mkdir()
-    cluster_pck = tmp_path / "cluster"
-    (cluster_pck / "agi_distributor").mkdir(parents=True)
-    (cluster_pck / "agi_distributor" / "cli.py").write_text("print('cli')\n", encoding="utf-8")
-
-    env = SimpleNamespace(
-        wenv_rel=Path("wenv/demo_worker"),
-        wenv_abs=tmp_path / "wenv" / "demo_worker",
-        active_app=app_path,
-        cluster_pck=cluster_pck,
-        envars={},
-        uv="uv",
-        export_local_bin="",
-        app="demo",
-        hw_rapids_capable=False,
-        is_local=lambda ip: True,
-    )
-    env.wenv_abs.mkdir(parents=True)
-    AGI.env = env
-    AGI._mode_auto = True
-    AGI._mode = AGI.DASK_MODE
-    AGI._workers = {"127.0.0.1": 1}
-    AGI._TIMEOUT = 1
-    AGI._worker_init_error = False
-
-    async def _fake_send_file(*_args, **_kwargs):
-        return None
-
-    async def _fake_kill(*_args, **_kwargs):
-        return None
-
-    async def _fake_wait_for_port_release(*_args, **_kwargs):
-        return True
-
-    async def _fake_detect_export_cmd(*_args, **_kwargs):
-        return ""
-
-    async def _fake_connect_scheduler_with_retry(*_args, **_kwargs):
-        raise ValueError("client boom")
-
-    original_sleep = asyncio.sleep
-    monkeypatch.setattr(AGI, "send_file", staticmethod(_fake_send_file))
-    monkeypatch.setattr(AGI, "_kill", staticmethod(_fake_kill))
-    monkeypatch.setattr(AGI, "_wait_for_port_release", staticmethod(_fake_wait_for_port_release))
-    monkeypatch.setattr(AGI, "_detect_export_cmd", staticmethod(_fake_detect_export_cmd))
-    monkeypatch.setattr(AGI, "_connect_scheduler_with_retry", staticmethod(_fake_connect_scheduler_with_retry))
-    monkeypatch.setattr(AGI, "_exec_bg", staticmethod(lambda *_args, **_kwargs: None))
-    monkeypatch.setattr(AGI, "_dask_env_prefix", staticmethod(lambda: ""))
-    monkeypatch.setattr(AGI, "_get_scheduler", staticmethod(lambda _scheduler: ("127.0.0.1", 8799)))
-    monkeypatch.setattr(agi_distributor_module.asyncio, "sleep", lambda *_args, **_kwargs: original_sleep(0))
-
-    with pytest.raises(RuntimeError, match="Failed to instantiate Dask Client"):
-        await AGI._start_scheduler("127.0.0.1")
 def test_train_capacity_missing_and_success(tmp_path):
     AGI._capacity_data_file = "capacity_data.csv"
     AGI._capacity_model_file = "capacity_model.pkl"
