+++ src/agilab/core/test/test_agi_distributor_runtime_misc_support.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 import asyncio
 import io
 import json
+import pickle
 import urllib.error
 from pathlib import Path
 from types import SimpleNamespace
@@ -187,3 +188,87 @@ def test_format_exception_chain_strips_generic_error_prefixes():
     text = runtime_misc_support.format_exception_chain(CustomError("CustomError: precise detail"))
 
     assert text.endswith("CustomError: precise detail")
+
+
+def test_load_capacity_predictor_returns_loaded_value(tmp_path):
+    model_path = tmp_path / "balancer_model.pkl"
+    model_path.write_bytes(b"pickle-bytes")
+
+    loaded = runtime_misc_support.load_capacity_predictor(
+        model_path,
+        load_fn=lambda stream: {"size": len(stream.read())},
+    )
+
+    assert loaded == {"size": len(b"pickle-bytes")}
+
+
+def test_load_capacity_predictor_retrains_when_missing(tmp_path):
+    calls = {"retrain": 0}
+
+    loaded = runtime_misc_support.load_capacity_predictor(
+        tmp_path / "missing.pkl",
+        retrain_fn=lambda: calls.__setitem__("retrain", calls["retrain"] + 1),
+    )
+
+    assert loaded is None
+    assert calls["retrain"] == 1
+
+
+def test_load_capacity_predictor_handles_legacy_module_error(tmp_path):
+    model_path = tmp_path / "balancer_model.pkl"
+    model_path.write_bytes(b"pickle-bytes")
+    calls = {"retrain": 0, "warnings": []}
+    log = SimpleNamespace(warning=lambda message, path, exc: calls["warnings"].append((message, path, str(exc))))
+
+    loaded = runtime_misc_support.load_capacity_predictor(
+        model_path,
+        load_fn=lambda _stream: (_ for _ in ()).throw(ModuleNotFoundError("numpy.core.numeric")),
+        retrain_fn=lambda: calls.__setitem__("retrain", calls["retrain"] + 1),
+        log=log,
+    )
+
+    assert loaded is None
+    assert calls["retrain"] == 1
+    assert calls["warnings"]
+    assert "numpy.core.numeric" in calls["warnings"][0][2]
+
+
+def test_hardware_supports_rapids_true_and_false(monkeypatch):
+    monkeypatch.setattr(runtime_misc_support.subprocess, "run", lambda *_a, **_k: None)
+    assert runtime_misc_support.hardware_supports_rapids() is True
+
+    monkeypatch.setattr(
+        runtime_misc_support.subprocess,
+        "run",
+        lambda *_a, **_k: (_ for _ in ()).throw(FileNotFoundError("nvidia-smi missing")),
+    )
+    assert runtime_misc_support.hardware_supports_rapids() is False
+
+
+def test_should_install_pip_checks_user_and_scripts_path(tmp_path):
+    assert runtime_misc_support.should_install_pip(
+        getuser_fn=lambda: "agi",
+        sys_prefix=str(tmp_path),
+    ) is False
+
+    assert runtime_misc_support.should_install_pip(
+        getuser_fn=lambda: "T01234",
+        sys_prefix=str(tmp_path),
+    ) is True
+
+    scripts_dir = tmp_path / "Scripts"
+    scripts_dir.mkdir()
+    (scripts_dir / "pip.exe").write_text("", encoding="utf-8")
+    assert runtime_misc_support.should_install_pip(
+        getuser_fn=lambda: "T01234",
+        sys_prefix=str(tmp_path),
+    ) is False
+
+
+def test_format_elapsed_uses_precisedelta_callback():
+    text = runtime_misc_support.format_elapsed(
+        12.5,
+        precisedelta_fn=lambda delta: f"{delta.total_seconds():.1f}s",
+    )
+
+    assert text == "12.5s"
